+++ packages/connect/src/connector/smoldot-light.ts
@@ -156,8 +156,26 @@ export const createScClient = (config?: Config): ScClient => {
       const internalChain = await client.addChain({
         chainSpec,
         potentialRelayChains: [...chains.values()],
-        jsonRpcCallback,
-      })
+        disableJsonRpc: jsonRpcCallback === undefined,
+      });
+
+      (async () => {
+        while (true) {
+          let jsonRpcResponse;
+          try {
+            jsonRpcResponse = await internalChain.nextJsonRpcResponse();
+          } catch(_) { break; }
+
+          // `nextJsonRpcResponse` throws an exception if we pass `disableJsonRpc: true` in the
+          // config. We pass `disableJsonRpc: true` if `jsonRpcCallback` is undefined. Therefore,
+          // this code is never reachable if `jsonRpcCallback` is undefined.
+          try {
+            jsonRpcCallback!(jsonRpcResponse)
+          } catch(error) {
+            console.error("JSON-RPC callback has thrown an exception:", error)
+          }
+        }
+      })()
 
       const chain: Chain = {
         sendJsonRpc: (rpc) => {